Modify Ideal->form function to have a better bijection.
Also added methods to determine if two definite forms are equivalent. The
method corresponding_form() was modified as follows: now the corresponding form
is N(ax - (b+cw)y)/N(I) instead of N(ax + (b+cw)y)/N(I). The previous version
was mathematically ok. Indeed, the two maps differ by the composition with the
inversion map on the group of Gaussian forms. The new choice of corresponding
form has the advantage that the map from forms to ideals and then back to forms
is exactly the identity at the level of classes (instead of the inversion map).